refactor sharedResources to shared/resources and shared/containers
