+++ src/react-apps/applications/runtime/src/sagas/index.ts
@@ -11,9 +11,9 @@ import FormResourceSagas from '../features/form/resources/sagas';
 import FormRulesSagas from '../features/form/rules/sagas';
 import FormValidationSagas from '../features/form/validation/sagas';
 import FormWorkflowSagas from '../features/form/workflow/sagas';
-import Attachments from '../sharedResources/attachments/attachmentSagas';
-import LanguageSagas from '../sharedResources/language/langaugeSagas';
-import ProfileSagas from '../sharedResources/profile/profileSagas';
+import Attachments from '../shared/resources/attachments/attachmentSagas';
+import LanguageSagas from '../shared/resources/language/langaugeSagas';
+import ProfileSagas from '../shared/resources/profile/profileSagas';
 
 function* root(): SagaIterator {
   yield fork(FormConfigSagas);
