+++ src/react-apps/applications/runtime/src/shared/resources/attachments/attachmentActionTypes.ts
