+++ src/react-apps/applications/runtime/src/shared/resources/attachments/attachmentReducer.ts
