+++ src/react-apps/applications/runtime/src/shared/resources/attachments/delete/deleteAttachmentActions.ts
