+++ src/react-apps/applications/runtime/src/shared/resources/language/fetch/fetchLanguageSagas.ts
@@ -1,6 +1,6 @@
 import { SagaIterator } from 'redux-saga';
 import { call, takeLatest } from 'redux-saga/effects';
-import { get } from '../../../utils/networking';
+import { get } from '../../../../utils/networking';
 import LanguageActions from './../languageActions';
 import { IFetchLanguage } from './fetchLanguageActions';
 import * as LanguageActionTypes from './fetchLanguageActionTypes';
