+++ src/react-apps/applications/runtime/src/shared/resources/language/languageActions.ts
@@ -1,5 +1,5 @@
 import { ActionCreatorsMapObject, bindActionCreators } from 'redux';
-import { store } from '../../store';
+import { store } from '../../../store';
 
 import * as FetchLanguage from './fetch/fetchLanguageActions';
