+++ src/react-apps/applications/runtime/src/shared/resources/profile/fetch/fetchProfileSagas.ts
@@ -1,6 +1,6 @@
 import { SagaIterator } from 'redux-saga';
 import { call, takeLatest } from 'redux-saga/effects';
-import { get } from '../../../../../shared/src/utils/networking';
+import { get } from '../../../../utils/networking';
 import ProfileActions from '../profileActions';
 import { IFetchProfile } from './fetchProfileActions';
 import * as ProfileActionTypes from './fetchProfileActionTypes';
