+++ src/react-apps/applications/runtime/src/shared/resources/utils/party.ts
