+++ src/react-apps/applications/runtime/src/types/index.ts
@@ -6,9 +6,9 @@ import { ILayoutState } from '../features/form/layout/reducer';
 import { IResourceState } from '../features/form/resources/reducer';
 import { IValidationState } from '../features/form/validation/reducer';
 import { IWorkflowState } from '../features/form/workflow/reducer';
-import { IAttachmentState } from '../sharedResources/attachments/attachmentReducer';
-import { ILanguageState } from '../sharedResources/language/languageReducers';
-import { IProfileState } from '../sharedResources/profile/profileReducers';
+import { IAttachmentState } from '../shared/resources/attachments/attachmentReducer';
+import { ILanguageState } from '../shared/resources/language/languageReducers';
+import { IProfileState } from '../shared/resources/profile/profileReducers';
 
 export interface IRuntimeState {
   formLayout: ILayoutState;
